==> code/Code_Bazzigher/Archiv_Codes/2_Correlations models.Rmd <==
---
title: "S2_Correlation models"
output:
  html_document: default
  pdf_document: default
date: "2024-05-08"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, results = 'hide', message = FALSE, warning = FALSE)
rm(list = ls())
```

# Set path Laura: ONLY USE FOR LAURA 
```{r}
base_path <- "//home.kt.ktzh.ch/B117T23$/Desktop/Riskktaking/Data"
```

# library
```{r}
library(ggplot2)
library(lme4)
library(dplyr)
library(lmerTest)
library(tidyverse)
library(maps)
library(ggrepel)
```

# Load preprocessed data countryfacts, GPS & WVS 
```{r}
GPS_data <- read.csv(file.path(base_path, "gps_cleaned.csv"), header=TRUE, as.is=TRUE)
WVS_data <- read.csv(file.path(base_path, "wvs_cleaned.csv"), header=TRUE, as.is=TRUE)
countryfacts <- read.csv(file.path(base_path, "countryfacts_cleaned.csv"), header=TRUE, as.is=TRUE)
head(GPS_data)
head(WVS_data)
head(countryfacts)
```

```{r}
str(GPS_data)
str(WVS_data)
str(countryfacts)
```


# Descriptive Statistics
## Frequency Distributions
```{r}
# Frequency distribution of countries in GPS_data dataset
table(GPS_data$country)

# Frequency distribution of genders in GPS_data dataset
table(GPS_data$gender)

# Frequency distribution of age groups in GPS_data dataset
table(cut(GPS_data$age, breaks=c(0, 20, 30, 40, 50, 60, 70, 80, 100)))

# Frequency distribution of countries in WVS_data dataset
table(WVS_data$country)

# Frequency distribution of genders in WVS_data dataset
table(WVS_data$gender)

# Frequency distribution of age groups in WVS_data dataset
table(cut(WVS_data$age, breaks=c(0, 20, 30, 40, 50, 60, 70, 80, 100)))
```

## Means and Medians
```{r}
# Mean and median of risk-taking in GPS_data dataset
mean(GPS_data$risktaking, na.rm = TRUE)
median(GPS_data$risktaking, na.rm = TRUE)

# Mean and median of age in GPS_data dataset
mean(GPS_data$age, na.rm = TRUE)
median(GPS_data$age, na.rm = TRUE)

# Mean and median of gender in GPS_data dataset
mean(GPS_data$gender, na.rm = TRUE)
median(GPS_data$gender, na.rm = TRUE)

# Mean and median of risk-taking in WVS_data dataset
mean(WVS_data$risktaking, na.rm = TRUE)
median(WVS_data$risktaking, na.rm = TRUE)

# Mean and median of age in WVS_data dataset
mean(WVS_data$age, na.rm = TRUE)
median(WVS_data$age, na.rm = TRUE)

# Mean and median of gender in WVS_data dataset
mean(WVS_data$gender, na.rm = TRUE)
median(WVS_data$gender, na.rm = TRUE)
```

## Standard Deviation and Variance
```{r}
# Standard deviation and variance of risk-taking in GPS_data dataset
sd(GPS_data$risktaking, na.rm = TRUE)
var(GPS_data$risktaking, na.rm = TRUE)

# Standard deviation and variance of age in GPS_data dataset
sd(GPS_data$age, na.rm = TRUE)
var(GPS_data$age, na.rm = TRUE)

# Standard deviation and variance of economic indicators in GPS_data dataset
sd(GPS_data$gender, na.rm = TRUE)
var(GPS_data$gender, na.rm = TRUE)

# Standard deviation and variance of risk-taking in WVS_data dataset
sd(WVS_data$risktaking, na.rm = TRUE)
var(WVS_data$risktaking, na.rm = TRUE)

# Standard deviation and variance of age in WVS_data dataset
sd(WVS_data$age, na.rm = TRUE)
var(WVS_data$age, na.rm = TRUE)

# Standard deviation and variance of gender in WVS_data dataset
sd(WVS_data$gender, na.rm = TRUE)
var(WVS_data$gender, na.rm = TRUE)
```

## Minimum and Maximum
```{r}
# Minimum and maximum of risk-taking in GPS_data dataset
min(GPS_data$risktaking, na.rm = TRUE)
max(GPS_data$risktaking, na.rm = TRUE)

# Minimum and maximum of age in GPS_data dataset
min(GPS_data$age, na.rm = TRUE)
max(GPS_data$age, na.rm = TRUE)

# Minimum and maximum of risk-taking in WVS_data dataset
min(WVS_data$risktaking, na.rm = TRUE)
max(WVS_data$risktaking, na.rm = TRUE)

# Minimum and maximum of age in WVS_data dataset
min(WVS_data$age, na.rm = TRUE)
max(WVS_data$age, na.rm = TRUE)
```

## Correlations
```{r}
# Correlations between numeric variables in GPS_data dataset
cor(GPS_data[, sapply(GPS_data, is.numeric)], use = "complete.obs")

# Correlations between numeric variables in WVS_data dataset
cor(WVS_data[, sapply(WVS_data, is.numeric)], use = "complete.obs")

# Correlations between numeric variables in countryfacts dataset
cor(countryfacts[, sapply(countryfacts, is.numeric)], use = "complete.obs")
```

## Average risk-taking
```{r}
# Average risk-taking by country in GPS_data dataset
aggregate(risktaking ~ country, data = GPS_data, FUN = mean, na.rm = TRUE)

# Average risk-taking by country in WVS_data dataset
aggregate(risktaking ~ country, data = WVS_data, FUN = mean, na.rm = TRUE)
```

## Gender-Based Analyses
```{r}
# Average risk-taking by gender in GPS_data dataset
aggregate(risktaking ~ gender, data = GPS_data, FUN = mean, na.rm = TRUE)

# Age distribution by gender in GPS_data dataset
aggregate(age ~ gender, data = GPS_data, FUN = summary)

# Average risk-taking by gender in WVS_data dataset
aggregate(risktaking ~ gender, data = WVS_data, FUN = mean, na.rm = TRUE)

# Age distribution by gender in WVS_data dataset
aggregate(age ~ gender, data = WVS_data, FUN = summary)
```


# Risk vs age with color-coded gender per Country
## GPS - Risk vs age with color-coded gender per Country 
```{r}
# Skalierung des Z-Scores für das Alter anpassen
GPS_data$age_scale <- 15 * GPS_data$age_scale + 42

# Risk vs age with color-coded gender per Country
ggplot(GPS_data, aes(age_scale, risktaking, color = factor(gender))) +
  geom_point(size = 0.1) +  
  geom_smooth(method = "lm") +
  geom_vline(xintercept = 42, linetype = "dashed", color = "black", size = 1) +  # Vertikale Linie für den Mittelwert
  scale_color_manual(values = c("blue", "red"), labels = c("Male", "Female")) +
  labs(color = "Gender", x = "Age", y = "Risk Taking") +  # Hier wurden die Achsentitel geändert
  scale_x_continuous(breaks = seq(0, 100, by = 15), limits = c(15, 100)) +  # Anpassung der Intervalle auf der X-Achse
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5))
```

## WVS - Risk vs age with color-coded gender per Country 
```{r}
WVS_data$age_scale <- 15 * WVS_data$age_scale + 42


ggplot(WVS_data, aes(age_scale, risktaking, color = as.factor(gender))) +
  geom_point(position = position_jitter(width = 0.1, height = 0.1), size = 0.1) +  
  geom_smooth(method = "lm") +
  geom_vline(xintercept = 42, linetype = "dashed", color = "black", size = 1) +  
  scale_color_manual(values = c("blue", "red"), labels = c("Male", "Female")) +
  labs(color = "Gender") +
  xlab("Age") +
  ylab("Risk Taking") +
  scale_x_continuous(breaks = seq(0, 100, by = 15), limits = c(15, 100)) +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5))
```


# Mixed-Models WVS
## mixed_models 0, 1 & 2
```{r}
# Clean data 
WVS_data <- na.omit(WVS_data)  # remove all NA's 
WVS_data <- WVS_data[complete.cases(WVS_data), ]

# Models 0, 1, & 2 
model0_wvs <- lmer(risktaking ~ 1 + (1|country), data = WVS_data)

## Model 1 wvs with age & gender
model1_wvs <- lmer(risktaking ~ 1 + age_scale + gender + (1 + age_scale + gender | country), 
                   data = WVS_data, 
                   control = lmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 30000)))
summary_model1_wvs <- summary(model1_wvs) 
print(summary_model1_wvs)

## Model 2 wvs with age, gender and hardship
model2_wvs <- lmer(risktaking ~ 1 + age_scale * hardship_index + 
                    gender * hardship_index + 
                    (1 + age_scale | country),
                data = WVS_data,
                control = lmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 30000)),
                REML = FALSE)
summary_model2_wvs <- summary(model2_wvs) 
print(summary_model2_wvs)

# ANOVA 
anova_results_wvs <- list(
  anova(model0_wvs, model1_wvs),
  anova(model1_wvs, model2_wvs)
)

print(anova_results_wvs)
```

# Mixed Models GPS
## mixed_models 0, 1 & 2
```{r}
# Clean data 
GPS_data <- na.omit(GPS_data)  # remove all NA's 
GPS_data <- GPS_data[complete.cases(GPS_data), ]

# Models 0, 1, & 2 
model0_gps <- lmer(risktaking ~ 1 + (1|country), data = GPS_data)

## Model 1 gps with age & gender
model1_gps <- lmer(risktaking ~ 1 + age_scale + gender + (1 + age_scale + gender | country), 
                   data = GPS_data, 
                   control = lmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 30000)))
summary_model1_gps <- summary(model1_gps) 
print(summary_model1_gps)

## Model 2 gps with age, gender and hardship
model2_gps <- lmer(risktaking ~ 1 + age_scale * hardship_index + 
                    gender * hardship_index + 
                    (1 + age_scale | country),
                data = GPS_data,
                control = lmerControl(optimizer = "bobyqa", optCtrl = list(maxfun = 30000)),
                REML = FALSE)
summary_model2_gps <- summary(model2_gps) 
print(summary_model2_gps)

# ANOVA
anova_results_gps <- list(
  anova(model0_gps, model1_gps),
  anova(model1_gps, model2_gps)
)

print(anova_results_gps)
```

```{r}
summary(model1_wvs)
summary(model2_wvs)
summary(model1_gps)
summary(model2_gps)
```

```{r}
# Extrahiere die Random Effects (Random Slopes und Intercepts) der Mixed Models für GPS und WVS
gps_random_effects <- coef(model1_gps)$country
wvs_random_effects <- coef(model1_wvs)$country

# Extrahiere die Länder (Zeilen) in beiden Modellen
gps_countries <- rownames(gps_random_effects)
wvs_countries <- rownames(wvs_random_effects)

# Finde die gemeinsamen Länder
common_countries <- intersect(gps_countries, wvs_countries)

# Filtere die Random Effects nur für die gemeinsamen Länder
gps_filtered <- gps_random_effects[common_countries, ]
wvs_filtered <- wvs_random_effects[common_countries, ]

# Zusammenführen der Ergebnisse in einem gemeinsamen DataFrame
new_data <- data.frame(
  isocode = common_countries,
  intercept_gps = gps_filtered[, "(Intercept)"],
  slope_age_gps = gps_filtered[, "age_scale"],
  slope_gender_gps = gps_filtered[, "gender"],
  intercept_WVS = wvs_filtered[, "(Intercept)"],
  slope_age_WVS = wvs_filtered[, "age_scale"],
  slope_gender_WVS = wvs_filtered[, "gender"]
)

# Füge die Ländercodes (isocode) aus countryfacts zu new_data hinzu
new_data <- new_data %>%
  left_join(countryfacts %>% select(country, code), by = c("isocode" = "country"))

# Überprüfe, ob die Ländercodes korrekt hinzugefügt wurden
head(new_data)

# Überprüfe die Struktur des neuen Datensatzes
head(new_data)
```

```{r}
head(new_data)
summary(new_data)

sum(is.na(new_data$intercept_gps))
sum(is.na(new_data$intercept_WVS))

```


```{r}
ggplot(new_data, aes(x = intercept_gps, y = intercept_WVS, label = code)) +
  geom_point(size = 1.5) +
  geom_text(aes(label = code), vjust = -0.5, hjust = -0.5) +
  geom_smooth(method = "lm", se = FALSE, linetype = "dashed") +
  labs(title = "Comparison of Intercept Values",
       x = "Intercept GPS",
       y = "Intercept WVS") +
  theme_minimal() +
  xlim(c(min(new_data$intercept_gps, na.rm = TRUE), max(new_data$intercept_gps, na.rm = TRUE))) +
  ylim(c(min(new_data$intercept_WVS, na.rm = TRUE), max(new_data$intercept_WVS, na.rm = TRUE))) +
  coord_fixed()

# Berechne die Korrelation der Intercepts
correlation_intercepts <- cor(new_data$intercept_gps, new_data$intercept_WVS)
print(correlation_intercepts)
```
```{r}
ggplot(new_data, aes(x = slope_age_gps, y = slope_age_WVS, label = code)) +
  geom_point(size = 1.5) +
  geom_text(aes(label = code), vjust = -0.5, hjust = -0.5) +
  geom_smooth(method = "lm", se = FALSE, linetype = "dashed") +
  labs(title = "Comparison of Age Slopes",
       x = "Age Slope GPS",
       y = "Age Slope WVS") +
  theme_minimal() +
  xlim(range(new_data$slope_age_gps, na.rm = TRUE)) +
  ylim(range(new_data$slope_age_WVS, na.rm = TRUE)) +
  coord_fixed()

correlation_intercepts <- cor(new_data$slope_age_gps, new_data$slope_age_WVS)
print(correlation_intercepts)
```

```{r}
ggplot(new_data, aes(x = slope_gender_gps, y = slope_gender_WVS, label = code)) +
  geom_point(size = 1.5) +
  geom_text(aes(label = code), vjust = -0.5, hjust = -0.5) +
  geom_smooth(method = "lm", se = FALSE, linetype = "dashed") +
  labs(title = "Comparison of Gender Slopes",
       x = "Gender Slope GPS",
       y = "Gender Slope WVS") +
  theme_minimal() +
  xlim(range(new_data$slope_gender_gps, na.rm = TRUE)) +
  ylim(range(new_data$slope_gender_WVS, na.rm = TRUE)) +
  coord_fixed()

correlation_intercepts <- cor(new_data$slope_gender_gps, new_data$slope_gender_WVS)
print(correlation_intercepts)
```
