==> code/Code_Bazzigher/Archiv_Codes/3_Specification_curve.Rmd <==
---
title: "3_Hardship_specification_curve"
output: html_document
date: "2024-05-08"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, results = 'hide', message = FALSE, warning = FALSE)
rm(list = ls())
```

# Set path Laura: ONLY USE FOR LAURA 
```{r}
base_path <- "//home.kt.ktzh.ch/B117T23$/Desktop/Riskktaking/Data"
```

# Library 
```{r}
library(tidyverse)
library(data.table)
library(tidyr)
library(maps)
library(haven)
library(ggplot2)
library(dplyr)
library(readxl)
library(lmerTest)
library(lme4)
library(broom)
library(tibble)
library(gridExtra)
```

# Load all data
```{r}
# Countryfacts
countryfacts <- read_excel(file.path(base_path, "Hardship_complete_2024.xlsx"))
str(countryfacts)

risktaking <- read.csv(file.path(base_path, "gps_cleaned.csv"), header=TRUE, as.is=TRUE)
str(risktaking)
```


```{r}
risktaking <- merge(countryfacts, risktaking, by = "code")
str(risktaking)
```

# Remove duplicated columns and unnecessary columns
```{r}
risktaking <- risktaking %>%
  select(-ends_with(".y")) %>%
  rename_with(~ gsub("\\.x", "", .), ends_with(".x"))

risktaking <- risktaking %>%
  select(-isocode, -ison, -label, -codeWVS, -age_scale)
str(risktaking)
```

# Define dependent variable
```{r}
dependent_var <- "risktaking"
```

# Define independent variables
```{r}
independent_vars <- c('homiciderate', 'gdp', 'infantmortality', 'gini', 'femalemale_primedu')
```

# Create all possible combinations of independent variables
```{r}
combinations <- unlist(lapply(1:length(independent_vars), 
                              function(x) combn(independent_vars, x, simplify = FALSE)), 
                       recursive = FALSE)
```

# Function to estimate a model and store the results
```{r}
estimate_model <- function(dependent_var, independent_vars, data) {
  formula <- as.formula(paste(dependent_var, "~", paste(independent_vars, collapse = "+")))
  model <- lm(formula, data)
  tidy_model <- tidy(model)
  return(tidy_model)
}
```

# Store results of all specifications
```{r}
results <- do.call(rbind, lapply(combinations, function(combo) {
  model_results <- estimate_model(dependent_var, combo, risktaking)
  model_results$variables <- paste(combo, collapse = ", ")
  model_results
}))

# Create a matrix of independent variables used in each specification
spec_matrix <- do.call(rbind, lapply(combinations, function(combo) {
  sapply(independent_vars, function(var) var %in% combo)
}))

# Convert to data frame for plotting
spec_matrix_df <- as.data.frame(spec_matrix)
colnames(spec_matrix_df) <- independent_vars
spec_matrix_df$specification <- seq_len(nrow(spec_matrix_df))

# Combine with results
results$specification <- seq_len(nrow(results))
results <- left_join(results, spec_matrix_df, by = "specification")

# Filter significant results
results$significant <- results$p.value < 0.05
```

# Create the Specification Curve Plot
```{r}
# Create the top plot (Specification Curve)
p1 <- ggplot(results, aes(x = specification, y = estimate, color = significant)) +
  geom_point(size = 1.5) +
  geom_line(aes(group = 1), size = 0.5) +
  scale_color_manual(values = c("TRUE" = "blue", "FALSE" = "red")) +
  labs(title = "Specification Curve Analysis", y = "Coefficient", x = "") +
  theme_minimal() +
  theme(axis.text.x = element_blank(), axis.ticks.x = element_blank(), 
        plot.margin = unit(c(1, 1, 0.5, 1), "cm"))

# Create the bottom plot (Specification Panel)
spec_matrix_long <- pivot_longer(spec_matrix_df, -specification, names_to = "variable", values_to = "included")

p2 <- ggplot(spec_matrix_long, aes(x = specification, y = variable, fill = included)) +
  geom_tile(color = "grey") +
  scale_fill_manual(values = c("TRUE" = "black", "FALSE" = "white")) +
  labs(y = "Independent Variables", x = "Model Specifications") +
  theme_minimal() +
  theme(axis.text.x = element_blank(), axis.ticks.x = element_blank(), 
        axis.text.y = element_text(size = 8),
        plot.margin = unit(c(0.5, 1, 1, 1), "cm"))

# Combine plots
grid.arrange(p1, p2, ncol = 1, heights = c(3, 1))
```
